change: migrate back to bpg/proxmox provider as its more feature rich
A note about cloud-init, the issue seems related to the initialize block
needing to tell the resource to re-create the cloud-init drive after
setting the user-data from my template. There are apparently also
issues with ide cloud-init drives and UEFI bios settings, but that seems
to be distro dependant. It didn't play nice with debian but was fine with
ubuntu.